Moved gradient function to render.h file
